requirements.txt
# Core IgniteFlow Dependencies
pyspark>=3.5.0
py4j>=0.10.9.7

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# MLOps & Experiment Tracking
mlflow>=2.5.0
boto3>=1.28.0
sagemaker>=2.175.0

# Configuration & Serialization
pyyaml>=6.0
jsonschema>=4.17.0
toml>=0.10.2

# Logging & Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-instrumentation>=0.40b0

# Data Quality & Validation
great-expectations>=0.17.0
cerberus>=1.3.4

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-spark>=0.6.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Development Tools
jupyter>=1.0.0
jupyterlab>=4.0.0
ipython>=8.14.0
notebook>=7.0.0

# Cloud Providers
# AWS
awscli>=2.13.0
s3fs>=2023.6.0

# Azure
azure-storage-blob>=12.17.0
azure-identity>=1.13.0
azure-keyvault-secrets>=4.7.0

# GCP
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0

# Kubernetes & Container Support
kubernetes>=27.2.0
docker>=6.1.0

# Streaming & Real-time Processing
kafka-python>=2.0.2
redis>=4.6.0
confluent-kafka>=2.2.0

# Database Connectors
psycopg2-binary>=2.9.0
pymongo>=4.4.0
sqlalchemy>=2.0.0
cx-Oracle>=8.3.0

# Utilities
requests>=2.31.0
click>=8.1.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pathlib2>=2.3.7

# Time & Date Handling
pendulum>=2.1.2
python-dateutil>=2.8.0

# Cryptography & Security
cryptography>=41.0.0
keyring>=24.2.0

# Performance & Profiling
memory-profiler>=0.61.0
line-profiler>=4.1.0
py-spy>=0.3.14

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0

# API & Web Services
fastapi>=0.101.0
uvicorn>=0.23.0
pydantic>=2.1.0

# Visualization (for demos and development)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development Dependencies (optional)
# Uncomment if needed for development
# jupytext>=1.15.0
# nbconvert>=7.7.0
# ipywidgets>=8.0.0